Use tasks.cr as alias for Lucky
The problem is that build-essentials is not available at run time, so we
can't build lucky tasks.

This build tasks.cr and outputs it as lucky. You use it the exact same
as Lucky CLI. The difference is you can't call init, or dev. But that
won't matter in Heroku.

This is also better because using the Lucky CLI will be super fast since
it is pre-built.